detect_in_video/requirements.txt
ffmpeg-python
pyTelegramBotAPI>=4.12.0
yt-dlp>=2023.3.4
loguru
requests